push login

+++ routes/web.php
@@ -1,10 +1,11 @@
 <?php
 
 use Illuminate\Support\Facades\Route;
-use App\Http\Controllers\StudentController;
+use App\Http\Controllers\AuthController;
 use App\Http\Controllers\ClassController;
-use App\Http\Controllers\ExtracurricularController;
+use App\Http\Controllers\StudentController;
 use App\Http\Controllers\TeacherController;
+use App\Http\Controllers\ExtracurricularController;
 
 /*
 |--------------------------------------------------------------------------
@@ -18,38 +19,43 @@ use App\Http\Controllers\TeacherController;
 */
 
 Route::get('/', function () {
-    return view('home', [
-        'name' => 'John Doe',
-        'role' => 'admin',
-        'buah' => ['pisang', 'apel', 'jeruk', 'kiwi', 'salak']
-    ]);
+
+    return view('home');
+    // return view('home', [
+    //     'name' => 'John Doe',
+    //     'role' => 'admin',
+    //     'buah' => ['pisang', 'apel', 'jeruk', 'kiwi', 'salak']
+    // ]);
 });
 
-Route::get('/students', [StudentController::class, 'index']);
-Route::get('/students/{id}', [StudentController::class, 'show']);
-Route::get('/student-add', [StudentController::class, 'create']);
-Route::post('/students', [StudentController::class, 'store']);
-Route::get('/student-edit/{id}', [StudentController::class, 'edit']);
-Route::put('/student/{id}', [StudentController::class, 'update']);
-Route::get('/student-delete/{id}', [StudentController::class, 'delete']);
-Route::delete('/student-destroy/{id}', [StudentController::class, 'destroy']);
-Route::get('/student-deleted', [StudentController::class, 'deletedStudent']);
-Route::get('/student/{id}/restore', [StudentController::class, 'restore']);
+Route::get('/login', [AuthController::class, 'login'])->name('login');
+Route::post('/login', [AuthController::class, 'authenticating']);
 
-Route::get('/class', [ClassController::class, 'index']);
-Route::get('/class-detail/{id}', [ClassController::class, 'show']);
-Route::get('/class-add', [ClassController::class, 'create']);
-Route::post('/class', [ClassController::class, 'store']);
+Route::get('/students', [StudentController::class, 'index'])->middleware('auth');
+Route::get('/students/{id}', [StudentController::class, 'show'])->middleware('auth');
+Route::get('/student-add', [StudentController::class, 'create'])->middleware('auth');
+Route::post('/students', [StudentController::class, 'store'])->middleware('auth');
+Route::get('/student-edit/{id}', [StudentController::class, 'edit'])->middleware('auth');
+Route::put('/student/{id}', [StudentController::class, 'update'])->middleware('auth');
+Route::get('/student-delete/{id}', [StudentController::class, 'delete'])->middleware('auth');
+Route::delete('/student-destroy/{id}', [StudentController::class, 'destroy'])->middleware('auth');
+Route::get('/student-deleted', [StudentController::class, 'deletedStudent'])->middleware('auth');
+Route::get('/student/{id}/restore', [StudentController::class, 'restore'])->middleware('auth');
 
-Route::get('/extracurricular', [ExtracurricularController::class, 'index']);
-Route::get('/extracurricular-detail/{id}', [ExtracurricularController::class, 'show']);
-Route::get('/extracurricular-add', [ExtracurricularController::class, 'create']);
-Route::post('/extracurricular', [ExtracurricularController::class, 'store']);
+Route::get('/class', [ClassController::class, 'index'])->middleware('auth');
+Route::get('/class-detail/{id}', [ClassController::class, 'show'])->middleware('auth');
+Route::get('/class-add', [ClassController::class, 'create'])->middleware('auth');
+Route::post('/class', [ClassController::class, 'store'])->middleware('auth');
 
-Route::get('/teacher', [TeacherController::class, 'index']);
-Route::get('/teacher-detail/{id}', [TeacherController::class, 'show']);
-Route::get('/teacher-add', [TeacherController::class, 'create']);
-Route::post('/teacher', [TeacherController::class, 'store']);
+Route::get('/extracurricular', [ExtracurricularController::class, 'index'])->middleware('auth');
+Route::get('/extracurricular-detail/{id}', [ExtracurricularController::class, 'show'])->middleware('auth');
+Route::get('/extracurricular-add', [ExtracurricularController::class, 'create'])->middleware('auth');
+Route::post('/extracurricular', [ExtracurricularController::class, 'store'])->middleware('auth');
+
+Route::get('/teacher', [TeacherController::class, 'index'])->middleware('auth');
+Route::get('/teacher-detail/{id}', [TeacherController::class, 'show'])->middleware('auth');
+Route::get('/teacher-add', [TeacherController::class, 'create'])->middleware('auth');
+Route::post('/teacher', [TeacherController::class, 'store'])->middleware('auth');
